룰렛 리스트 수정 추가, id 타입변경, ui 일부 수정
룰렛 리스트 수정 추가, id 타입변경, ui 일부 수정

+++ data/src/main/java/com/example/data/dao/RouletteDao.kt
@@ -20,5 +20,5 @@ interface RouletteDao {
     fun update(roulette: RouletteEntity)
 
     @Query("DELETE FROM Roulette WHERE id = :id")
-    fun delete(id: Int)
+    fun delete(id: Long)
 }
